src/http_functions/requirements.txt
azure-functions
pandas
aiohttp  
azure-identity  
azure-storage-blob